Deprecated functions from skimr v1

Skimr used to offer functions that combined skimming with a secondary effect, like reshaping the data, building a list or printing the results. Some of these behaviors are no longer necessary. skim() always returns a wide data frame. Others have been replaced by functions that do a single thing. partition() creates a list-like object from a skimmed data frame.

skim_to_wide(.data, ...)

skim_to_list(.data, ...)

skim_format(...)

Arguments

.data

A tibble, or an object that can be coerced into a tibble.

...

Columns to select for skimming. When none are provided, the default is to skim all columns.

Value

Either A skim_df or a skim_list object.

Functions

skim_to_wide: skim() always produces a wide data frame.

skim_to_list: partition() creates a list.

skim_format: print() and skim_with() set options.