Only show a subset of summary statistics after skimming

This function is a variant of dplyr::select() designed to work with skim_df objects. When using focus(), skimr metadata columns are kept, and skimr print methods are still utilized. Otherwise, the signature and behavior is identical to dplyr::select().

focus(.data, ...)

Arguments

.data
A skim_df object.

...
One or more unquoted expressions separated by commas. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables.

Examples

# Compare
iris %>%
  skim() %>%
  dplyr::select(n_missing)

#> # A tibble: 5 x 1
#>   n_missing
#>       <int>
#> 1         0
#> 2         0
#> 3         0
#> 4         0
#> 5         0

iris %>%
  skim() %>%
  focus(n_missing)

#> ── Data Summary ────────────────────────
#>                            Values    
#> Name                       Piped data
#> Number of rows             150       
#> Number of columns          5         
#> _______________________              
#> Column type frequency:               
#>   factor                   1         
#>   numeric                  4         
#> ________________________             
#> Group variables            None      
#> 
#> ── Variable type: factor ───────────────────────────────────────────────────────
#> 
#> 
#> ── Variable type: numeric ──────────────────────────────────────────────────────
#> 

# This is equivalent to
iris %>%
  skim() %>%
  dplyr::select(skim_variable, skim_type, n_missing)

#> ── Data Summary ────────────────────────
#>                            Values    
#> Name                       Piped data
#> Number of rows             150       
#> Number of columns          5         
#> _______________________              
#> Column type frequency:               
#>   factor                   1         
#>   numeric                  4         
#> ________________________             
#> Group variables            None      
#> 
#> ── Variable type: factor ───────────────────────────────────────────────────────
#> 
#> 
#> ── Variable type: numeric ──────────────────────────────────────────────────────
#>