Skim a data frame, getting useful summary statistics

skim() is an alternative to summary(), quickly providing a broad overview of a data frame. It handles data of all types, dispatching a different set of summary functions based on the types of columns in the data frame.

skim(.data, ...)

skim_tee(.data, ...)

Arguments

.data
A tibble, or an object that can be coerced into a tibble.

...
Additional options, normally used to list individual unquoted column names.

Value

A skim_df object, which can be treated like a tibble in most instances.

Details

Each call produces a skim_df, which is a fundamentally a tibble with a special print method. Instead of showing the result in a long format, skim prints a wide version of your data with formatting applied to each column. Printing does not change the structure of the skim_df, which remains a long tibble.

If you just want to see the printed output, call skim_tee() instead. This function returns the original data frame.

If you want to work with a data frame that resembles the printed output, call skim_to_wide() or for a named list of data frames by type skim_to_list(). Note that all of the columns in the data frames produced by these functions are character. The intent is that you will be processing the printed result further, not the original data.

skim() is designed to operate in pipes and to generally play nicely with other tidyverse functions. This means that you can use tidyselect helpers within skim to select or drop specific columns for summary. You can also further work with a skim_df using dplyr functions in a pipeline.

Customizing skim

skim() is an intentionally simple function, with minimal arguments like summary(). Nonetheless, this package provides two broad approaches to how you can customize skim()'s behavior. You can customize the functions that are called to produce summary statistics with skim_with(). You can customize how the output is displayed with skim_format().

Unicode rendering

If the rendered examples show unencoded values such as <U+2587> you will need to change your locale to allow proper rendering. Please review the Using Skimr vignette for more information (vignette("Using_skimr", package = "skimr")).

Examples

skim(iris)

#> Skim summary statistics
#>  n obs: 150 
#>  n variables: 5 
#> 
#> ── Variable type:factor ────────────────────────────────────────────────────────
#>  variable missing complete   n n_unique                       top_counts
#>   Species       0      150 150        3 set: 50, ver: 50, vir: 50, NA: 0
#>  ordered
#>    FALSE
#> 
#> ── Variable type:numeric ───────────────────────────────────────────────────────
#>      variable missing complete   n mean   sd  p0 p25  p50 p75 p100     hist
#>  Petal.Length       0      150 150 3.76 1.77 1   1.6 4.35 5.1  6.9 ▇▁▁▂▅▅▃▁
#>   Petal.Width       0      150 150 1.2  0.76 0.1 0.3 1.3  1.8  2.5 ▇▁▁▅▃▃▂▂
#>  Sepal.Length       0      150 150 5.84 0.83 4.3 5.1 5.8  6.4  7.9 ▂▇▅▇▆▅▂▂
#>   Sepal.Width       0      150 150 3.06 0.44 2   2.8 3    3.3  4.4 ▁▂▅▇▃▂▁▁

# Use tidyselect
skim(iris, Species)

#> Skim summary statistics
#>  n obs: 150 
#>  n variables: 5 
#> 
#> ── Variable type:factor ────────────────────────────────────────────────────────
#>  variable missing complete   n n_unique                       top_counts
#>   Species       0      150 150        3 set: 50, ver: 50, vir: 50, NA: 0
#>  ordered
#>    FALSE

skim(iris, starts_with("Sepal"))

#> Skim summary statistics
#>  n obs: 150 
#>  n variables: 5 
#> 
#> ── Variable type:numeric ───────────────────────────────────────────────────────
#>      variable missing complete   n mean   sd  p0 p25 p50 p75 p100     hist
#>  Sepal.Length       0      150 150 5.84 0.83 4.3 5.1 5.8 6.4  7.9 ▂▇▅▇▆▅▂▂
#>   Sepal.Width       0      150 150 3.06 0.44 2   2.8 3   3.3  4.4 ▁▂▅▇▃▂▁▁

# Skim also works groupwise
dplyr::group_by(iris, Species) %>% skim()

#> Skim summary statistics
#>  n obs: 150 
#>  n variables: 5 
#>  group variables: Species 
#> 
#> ── Variable type:numeric ───────────────────────────────────────────────────────
#>     Species     variable missing complete  n mean   sd  p0  p25  p50  p75 p100
#>      setosa Petal.Length       0       50 50 1.46 0.17 1   1.4  1.5  1.58  1.9
#>      setosa  Petal.Width       0       50 50 0.25 0.11 0.1 0.2  0.2  0.3   0.6
#>      setosa Sepal.Length       0       50 50 5.01 0.35 4.3 4.8  5    5.2   5.8
#>      setosa  Sepal.Width       0       50 50 3.43 0.38 2.3 3.2  3.4  3.68  4.4
#>  versicolor Petal.Length       0       50 50 4.26 0.47 3   4    4.35 4.6   5.1
#>  versicolor  Petal.Width       0       50 50 1.33 0.2  1   1.2  1.3  1.5   1.8
#>  versicolor Sepal.Length       0       50 50 5.94 0.52 4.9 5.6  5.9  6.3   7  
#>  versicolor  Sepal.Width       0       50 50 2.77 0.31 2   2.52 2.8  3     3.4
#>   virginica Petal.Length       0       50 50 5.55 0.55 4.5 5.1  5.55 5.88  6.9
#>   virginica  Petal.Width       0       50 50 2.03 0.27 1.4 1.8  2    2.3   2.5
#>   virginica Sepal.Length       0       50 50 6.59 0.64 4.9 6.23 6.5  6.9   7.9
#>   virginica  Sepal.Width       0       50 50 2.97 0.32 2.2 2.8  3    3.18  3.8
#>      hist
#>  ▁▁▅▇▇▅▂▁
#>  ▂▇▁▂▂▁▁▁
#>  ▂▃▅▇▇▃▁▂
#>  ▁▁▃▅▇▃▂▁
#>  ▁▃▂▆▆▇▇▃
#>  ▆▃▇▅▆▂▁▁
#>  ▃▂▇▇▇▃▅▂
#>  ▁▂▃▅▃▇▃▁
#>  ▂▇▃▇▅▂▁▂
#>  ▂▁▇▃▃▆▅▃
#>  ▁▁▃▇▅▃▂▃
#>  ▁▃▇▇▅▃▁▂

# Skim pipelines; now we work with the tall format
skim(iris) %>% as.data.frame()

#>        variable    type       stat      level       value formatted
#> 1  Sepal.Length numeric    missing       .all   0.0000000         0
#> 2  Sepal.Length numeric   complete       .all 150.0000000       150
#> 3  Sepal.Length numeric          n       .all 150.0000000       150
#> 4  Sepal.Length numeric       mean       .all   5.8433333      5.84
#> 5  Sepal.Length numeric         sd       .all   0.8280661      0.83
#> 6  Sepal.Length numeric         p0       .all   4.3000000       4.3
#> 7  Sepal.Length numeric        p25       .all   5.1000000       5.1
#> 8  Sepal.Length numeric        p50       .all   5.8000000       5.8
#> 9  Sepal.Length numeric        p75       .all   6.4000000       6.4
#> 10 Sepal.Length numeric       p100       .all   7.9000000       7.9
#> 11 Sepal.Length numeric       hist       .all          NA  ▂▇▅▇▆▅▂▂
#> 12  Sepal.Width numeric    missing       .all   0.0000000         0
#> 13  Sepal.Width numeric   complete       .all 150.0000000       150
#> 14  Sepal.Width numeric          n       .all 150.0000000       150
#> 15  Sepal.Width numeric       mean       .all   3.0573333      3.06
#> 16  Sepal.Width numeric         sd       .all   0.4358663      0.44
#> 17  Sepal.Width numeric         p0       .all   2.0000000         2
#> 18  Sepal.Width numeric        p25       .all   2.8000000       2.8
#> 19  Sepal.Width numeric        p50       .all   3.0000000         3
#> 20  Sepal.Width numeric        p75       .all   3.3000000       3.3
#> 21  Sepal.Width numeric       p100       .all   4.4000000       4.4
#> 22  Sepal.Width numeric       hist       .all          NA  ▁▂▅▇▃▂▁▁
#> 23 Petal.Length numeric    missing       .all   0.0000000         0
#> 24 Petal.Length numeric   complete       .all 150.0000000       150
#> 25 Petal.Length numeric          n       .all 150.0000000       150
#> 26 Petal.Length numeric       mean       .all   3.7580000      3.76
#> 27 Petal.Length numeric         sd       .all   1.7652982      1.77
#> 28 Petal.Length numeric         p0       .all   1.0000000         1
#> 29 Petal.Length numeric        p25       .all   1.6000000       1.6
#> 30 Petal.Length numeric        p50       .all   4.3500000      4.35
#> 31 Petal.Length numeric        p75       .all   5.1000000       5.1
#> 32 Petal.Length numeric       p100       .all   6.9000000       6.9
#> 33 Petal.Length numeric       hist       .all          NA  ▇▁▁▂▅▅▃▁
#> 34  Petal.Width numeric    missing       .all   0.0000000         0
#> 35  Petal.Width numeric   complete       .all 150.0000000       150
#> 36  Petal.Width numeric          n       .all 150.0000000       150
#> 37  Petal.Width numeric       mean       .all   1.1993333       1.2
#> 38  Petal.Width numeric         sd       .all   0.7622377      0.76
#> 39  Petal.Width numeric         p0       .all   0.1000000       0.1
#> 40  Petal.Width numeric        p25       .all   0.3000000       0.3
#> 41  Petal.Width numeric        p50       .all   1.3000000       1.3
#> 42  Petal.Width numeric        p75       .all   1.8000000       1.8
#> 43  Petal.Width numeric       p100       .all   2.5000000       2.5
#> 44  Petal.Width numeric       hist       .all          NA  ▇▁▁▅▃▃▂▂
#> 45      Species  factor    missing       .all   0.0000000         0
#> 46      Species  factor   complete       .all 150.0000000       150
#> 47      Species  factor          n       .all 150.0000000       150
#> 48      Species  factor   n_unique       .all   3.0000000         3
#> 49      Species  factor top_counts     setosa  50.0000000   set: 50
#> 50      Species  factor top_counts versicolor  50.0000000   ver: 50
#> 51      Species  factor top_counts  virginica  50.0000000   vir: 50
#> 52      Species  factor top_counts       <NA>   0.0000000     NA: 0
#> 53      Species  factor    ordered       .all   0.0000000     FALSE

skim(iris) %>% dplyr::filter(type == "factor")

#> # A tibble: 9 x 6
#>   variable type   stat       level      value formatted
#>   <chr>    <chr>  <chr>      <chr>      <dbl> <chr>    
#> 1 Species  factor missing    .all           0 0        
#> 2 Species  factor complete   .all         150 150      
#> 3 Species  factor n          .all         150 150      
#> 4 Species  factor n_unique   .all           3 3        
#> 5 Species  factor top_counts setosa        50 set: 50  
#> 6 Species  factor top_counts versicolor    50 ver: 50  
#> 7 Species  factor top_counts virginica     50 vir: 50  
#> 8 Species  factor top_counts <NA>           0 NA: 0    
#> 9 Species  factor ordered    .all           0 FALSE    

# Which column as the greatest mean value?
skim(iris) %>%
  dplyr::filter(stat == "mean") %>%
  dplyr::arrange(dplyr::desc(value))

#> # A tibble: 4 x 6
#>   variable     type    stat  level value formatted
#>   <chr>        <chr>   <chr> <chr> <dbl> <chr>    
#> 1 Sepal.Length numeric mean  .all   5.84 5.84     
#> 2 Petal.Length numeric mean  .all   3.76 3.76     
#> 3 Sepal.Width  numeric mean  .all   3.06 3.06     
#> 4 Petal.Width  numeric mean  .all   1.20 1.2      

# Use skim_tee to view the skim results and
# continue using the original data.
chickwts %>% skim_tee() %>% dplyr::filter(feed == "sunflower")

#> Skim summary statistics
#>  n obs: 71 
#>  n variables: 2 
#> 
#> ── Variable type:factor ────────────────────────────────────────────────────────
#>  variable missing complete  n n_unique                         top_counts
#>      feed       0       71 71        6 soy: 14, cas: 12, lin: 12, sun: 12
#>  ordered
#>    FALSE
#> 
#> ── Variable type:numeric ───────────────────────────────────────────────────────
#>  variable missing complete  n   mean    sd  p0   p25 p50   p75 p100     hist
#>    weight       0       71 71 261.31 78.07 108 204.5 258 323.5  423 ▃▅▅▇▃▇▂▂

#>    weight      feed
#> 1     423 sunflower
#> 2     340 sunflower
#> 3     392 sunflower
#> 4     339 sunflower
#> 5     341 sunflower
#> 6     226 sunflower
#> 7     320 sunflower
#> 8     295 sunflower
#> 9     334 sunflower
#> 10    322 sunflower
#> 11    297 sunflower
#> 12    318 sunflower