Create "long" skim output

Skim results returned as a tidy long data frame with four columns: variable, type, stat and formatted.

to_long(.data, ..., skim_fun = skim)

# S3 method for default
to_long(.data, ..., skim_fun = skim)

# S3 method for skim_df
to_long(.data, ..., skim_fun = skim)

Arguments

.data
A data frame or an object that can be coerced into a data frame.

...
Columns to select for skimming. When none are provided, the default is to skim all columns.

skim_fun
The skim function used.

Value

A tibble

Methods (by class)

default: Skim a data frame and convert the results to a long data frame.

skim_df: Transform a skim_df to a long data frame.

Examples

to_long(iris)

#> # A tibble: 45 x 4
#>    skim_type skim_variable stat          formatted
#>    <chr>     <chr>         <chr>         <chr>    
#>  1 factor    Species       n_missing     0        
#>  2 numeric   Sepal.Length  n_missing     0        
#>  3 numeric   Sepal.Width   n_missing     0        
#>  4 numeric   Petal.Length  n_missing     0        
#>  5 numeric   Petal.Width   n_missing     0        
#>  6 factor    Species       complete_rate 1        
#>  7 numeric   Sepal.Length  complete_rate 1        
#>  8 numeric   Sepal.Width   complete_rate 1        
#>  9 numeric   Petal.Length  complete_rate 1        
#> 10 numeric   Petal.Width   complete_rate 1        
#> # … with 35 more rows

to_long(skim(iris))

#> # A tibble: 45 x 4
#>    skim_type skim_variable stat          formatted
#>    <chr>     <chr>         <chr>         <chr>    
#>  1 factor    Species       n_missing     0        
#>  2 numeric   Sepal.Length  n_missing     0        
#>  3 numeric   Sepal.Width   n_missing     0        
#>  4 numeric   Petal.Length  n_missing     0        
#>  5 numeric   Petal.Width   n_missing     0        
#>  6 factor    Species       complete_rate 1        
#>  7 numeric   Sepal.Length  complete_rate 1        
#>  8 numeric   Sepal.Width   complete_rate 1        
#>  9 numeric   Petal.Length  complete_rate 1        
#> 10 numeric   Petal.Width   complete_rate 1        
#> # … with 35 more rows