Working with skimr's printed output

These functions provide two approaches for handling the wide format produced when you print skim_df. skim_to_wide() returns a wide data frame with one row per variable and NA for statistics not calculated for a given type. skim_to_list() creates a list of wide tibbles, one for each type of vector within your data frame.

skim_to_wide(x, ...)

skim_to_list(x, ...)

Arguments

x
A data frame.

...
Further arguments passed to or from other methods.

Value

A wide data frame or a list of wide data frames.

Details

Note that in both cases, all columns are character vectors. This gives you additional control of the printed output, but not the original data.

Examples

# Treat the printed output as a wide data frame
skim_to_wide(iris)

#> # A tibble: 5 x 16
#>   type  variable missing complete n     n_unique top_counts ordered mean  sd   
#>   <chr> <chr>    <chr>   <chr>    <chr> <chr>    <chr>      <chr>   <chr> <chr>
#> 1 fact… Species  0       150      150   3        set: 50, … FALSE   <NA>  <NA> 
#> 2 nume… Petal.L… 0       150      150   <NA>     <NA>       <NA>    3.76  1.77 
#> 3 nume… Petal.W… 0       150      150   <NA>     <NA>       <NA>    "1.2… 0.76 
#> 4 nume… Sepal.L… 0       150      150   <NA>     <NA>       <NA>    5.84  0.83 
#> 5 nume… Sepal.W… 0       150      150   <NA>     <NA>       <NA>    3.06  0.44 
#> # … with 6 more variables: p0 <chr>, p25 <chr>, p50 <chr>, p75 <chr>,
#> #   p100 <chr>, hist <chr>

iris %>% skim_to_wide()

#> # A tibble: 5 x 16
#>   type  variable missing complete n     n_unique top_counts ordered mean  sd   
#>   <chr> <chr>    <chr>   <chr>    <chr> <chr>    <chr>      <chr>   <chr> <chr>
#> 1 fact… Species  0       150      150   3        set: 50, … FALSE   <NA>  <NA> 
#> 2 nume… Petal.L… 0       150      150   <NA>     <NA>       <NA>    3.76  1.77 
#> 3 nume… Petal.W… 0       150      150   <NA>     <NA>       <NA>    "1.2… 0.76 
#> 4 nume… Sepal.L… 0       150      150   <NA>     <NA>       <NA>    5.84  0.83 
#> 5 nume… Sepal.W… 0       150      150   <NA>     <NA>       <NA>    3.06  0.44 
#> # … with 6 more variables: p0 <chr>, p25 <chr>, p50 <chr>, p75 <chr>,
#> #   p100 <chr>, hist <chr>

iris %>%
  skim_to_wide() %>%
  dplyr::filter(type == "factor") %>%
  dplyr::select(top_counts)

#> # A tibble: 1 x 1
#>   top_counts                      
#>   <chr>                           
#> 1 set: 50, ver: 50, vir: 50, NA: 0

# Treat the printed output as a list of data frames
skim_to_list(iris)

#> $factor
#> # A tibble: 1 x 7
#>   variable missing complete n     n_unique top_counts                    ordered
#> * <chr>    <chr>   <chr>    <chr> <chr>    <chr>                         <chr>  
#> 1 Species  0       150      150   3        set: 50, ver: 50, vir: 50, N… FALSE  
#> 
#> $numeric
#> # A tibble: 4 x 12
#>   variable missing complete n     mean  sd    p0    p25   p50   p75   p100 
#> * <chr>    <chr>   <chr>    <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1 Petal.L… 0       150      150   3.76  1.77  "1  " 1.6   4.35  5.1   6.9  
#> 2 Petal.W… 0       150      150   "1.2… 0.76  0.1   0.3   "1.3… 1.8   2.5  
#> 3 Sepal.L… 0       150      150   5.84  0.83  4.3   5.1   "5.8… 6.4   7.9  
#> 4 Sepal.W… 0       150      150   3.06  0.44  "2  " 2.8   "3  … 3.3   4.4  
#> # … with 1 more variable: hist <chr>
#> 

iris %>% skim_to_list()

#> $factor
#> # A tibble: 1 x 7
#>   variable missing complete n     n_unique top_counts                    ordered
#> * <chr>    <chr>   <chr>    <chr> <chr>    <chr>                         <chr>  
#> 1 Species  0       150      150   3        set: 50, ver: 50, vir: 50, N… FALSE  
#> 
#> $numeric
#> # A tibble: 4 x 12
#>   variable missing complete n     mean  sd    p0    p25   p50   p75   p100 
#> * <chr>    <chr>   <chr>    <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1 Petal.L… 0       150      150   3.76  1.77  "1  " 1.6   4.35  5.1   6.9  
#> 2 Petal.W… 0       150      150   "1.2… 0.76  0.1   0.3   "1.3… 1.8   2.5  
#> 3 Sepal.L… 0       150      150   5.84  0.83  4.3   5.1   "5.8… 6.4   7.9  
#> 4 Sepal.W… 0       150      150   3.06  0.44  "2  " 2.8   "3  … 3.3   4.4  
#> # … with 1 more variable: hist <chr>
#> 

# Save the result
sl <- iris %>% skim_to_list()
sl[["numeric"]]

#> # A tibble: 4 x 12
#>   variable missing complete n     mean  sd    p0    p25   p50   p75   p100 
#> * <chr>    <chr>   <chr>    <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1 Petal.L… 0       150      150   3.76  1.77  "1  " 1.6   4.35  5.1   6.9  
#> 2 Petal.W… 0       150      150   "1.2… 0.76  0.1   0.3   "1.3… 1.8   2.5  
#> 3 Sepal.L… 0       150      150   5.84  0.83  4.3   5.1   "5.8… 6.4   7.9  
#> 4 Sepal.W… 0       150      150   3.06  0.44  "2  " 2.8   "3  … 3.3   4.4  
#> # … with 1 more variable: hist <chr>

kable(sl$numeric)

#> 
#> 
#> |   variable   | missing | complete |  n  | mean |  sd  | p0  | p25 | p50  | p75 | p100 |   hist   |
#> |--------------|---------|----------|-----|------|------|-----|-----|------|-----|------|----------|
#> | Petal.Length |    0    |   150    | 150 | 3.76 | 1.77 |  1  | 1.6 | 4.35 | 5.1 | 6.9  | ▇▁▁▂▅▅▃▁ |
#> | Petal.Width  |    0    |   150    | 150 | 1.2  | 0.76 | 0.1 | 0.3 | 1.3  | 1.8 | 2.5  | ▇▁▁▅▃▃▂▂ |
#> | Sepal.Length |    0    |   150    | 150 | 5.84 | 0.83 | 4.3 | 5.1 | 5.8  | 6.4 | 7.9  | ▂▇▅▇▆▅▂▂ |
#> | Sepal.Width  |    0    |   150    | 150 | 3.06 | 0.44 |  2  | 2.8 |  3   | 3.3 | 4.4  | ▁▂▅▇▃▂▁▁ |

# Or grouped, this uses the magrittr exposition pipe
# see ?magrittr::`%$%`
library(magrittr)

#> 
#> Attaching package: ‘magrittr’

#> The following objects are masked from ‘package:testthat’:
#> 
#>     equals, is_less_than, not

iris %>%
  dplyr::group_by(Species) %>%
  skim_to_list() %$%
  kable(numeric)

#> 
#> 
#> |  Species   |   variable   | missing | complete | n  | mean |  sd  | p0  | p25  | p50  | p75  | p100 |   hist   |
#> |------------|--------------|---------|----------|----|------|------|-----|------|------|------|------|----------|
#> |   setosa   | Petal.Length |    0    |    50    | 50 | 1.46 | 0.17 |  1  | 1.4  | 1.5  | 1.58 | 1.9  | ▁▁▅▇▇▅▂▁ |
#> |   setosa   | Petal.Width  |    0    |    50    | 50 | 0.25 | 0.11 | 0.1 | 0.2  | 0.2  | 0.3  | 0.6  | ▂▇▁▂▂▁▁▁ |
#> |   setosa   | Sepal.Length |    0    |    50    | 50 | 5.01 | 0.35 | 4.3 | 4.8  |  5   | 5.2  | 5.8  | ▂▃▅▇▇▃▁▂ |
#> |   setosa   | Sepal.Width  |    0    |    50    | 50 | 3.43 | 0.38 | 2.3 | 3.2  | 3.4  | 3.68 | 4.4  | ▁▁▃▅▇▃▂▁ |
#> | versicolor | Petal.Length |    0    |    50    | 50 | 4.26 | 0.47 |  3  |  4   | 4.35 | 4.6  | 5.1  | ▁▃▂▆▆▇▇▃ |
#> | versicolor | Petal.Width  |    0    |    50    | 50 | 1.33 | 0.2  |  1  | 1.2  | 1.3  | 1.5  | 1.8  | ▆▃▇▅▆▂▁▁ |
#> | versicolor | Sepal.Length |    0    |    50    | 50 | 5.94 | 0.52 | 4.9 | 5.6  | 5.9  | 6.3  |  7   | ▃▂▇▇▇▃▅▂ |
#> | versicolor | Sepal.Width  |    0    |    50    | 50 | 2.77 | 0.31 |  2  | 2.52 | 2.8  |  3   | 3.4  | ▁▂▃▅▃▇▃▁ |
#> | virginica  | Petal.Length |    0    |    50    | 50 | 5.55 | 0.55 | 4.5 | 5.1  | 5.55 | 5.88 | 6.9  | ▂▇▃▇▅▂▁▂ |
#> | virginica  | Petal.Width  |    0    |    50    | 50 | 2.03 | 0.27 | 1.4 | 1.8  |  2   | 2.3  | 2.5  | ▂▁▇▃▃▆▅▃ |
#> | virginica  | Sepal.Length |    0    |    50    | 50 | 6.59 | 0.64 | 4.9 | 6.23 | 6.5  | 6.9  | 7.9  | ▁▁▃▇▅▃▂▃ |
#> | virginica  | Sepal.Width  |    0    |    50    | 50 | 2.97 | 0.32 | 2.2 | 2.8  |  3   | 3.18 | 3.8  | ▁▃▇▇▅▃▁▂ |