Use dplyr verb mutate on skim_df objects.

Use dplyr verb mutate on skim_df objects.

# S3 method for skim_df
mutate(.data, ...)

Arguments

.data
A skim object

...
Name-value pairs of expressions. The name of each argument will be the name of a new variable, and the value will be its corresponding value.

Value

skim_df object coerced to a data frame.

See also