Print skim objects

skimr has custom print methods for all supported objects. Default printing methods for knitr/ rmarkdown documents is also provided.

Usage

# S3 method for skim_df
print(
  x,
  include_summary = TRUE,
  n = Inf,
  width = Inf,
  summary_rule_width = getOption("skimr_summary_rule_width", default = 40),
  ...
)

# S3 method for skim_list
print(x, n = Inf, width = Inf, ...)

# S3 method for summary_skim_df
print(x, .summary_rule_width = 40, ...)

Arguments

x

Object to format or print.

include_summary

Whether a summary of the data frame should be printed

n

Number of rows to show. If NULL, the default, will print all rows if less than the print_max option. Otherwise, will print as many rows as specified by the print_min option.

width

Width of text output to generate. This defaults to NULL, which means use the width option.

summary_rule_width

Width of Data Summary cli rule, defaults to 40.

...

Passed on to tbl_format_setup().

.summary_rule_width

the width for the main rule above the summary.

Methods (by class)

print(skim_df): Print a skimmed data frame (skim_df from skim()).

print(skim_list): Print a skim_list, a list of skim_df objects.

print(summary_skim_df): Print method for a summary_skim_df object.

Printing options

For better or for worse, skimr often produces more output than can fit in the standard R console. Fortunately, most modern environments like RStudio and Jupyter support more than 80 character outputs. Call options(width = 90) to get a better experience with skimr.

The print methods in skimr wrap those in the tibble package. You can control printing behavior using the same global options.

Behavior in dplyr pipelines

Printing a skim_df requires specific columns that might be dropped when using dplyr::select() or dplyr::summarize() on a skim_df. In those cases, this method falls back to tibble::print.tbl().

Options for controlling print behavior

You can control the width rule line for the printed subtables with an option: skimr_table_header_width.

See also